refactor(effect): also split error/ + discriminants/ subfolders, mirror docs
Completes the effect/ family reorg started by the v0.0.9 batch. The
original 5 rules (effect-error-erasure, effect-promise, no-effect-error-
coalescing, either-discriminant, tag-discriminant) move into two more
sub-families so the parent effect/index.ts is a pure facade — only
imports from sub-barrels, no flat rule imports.

- effect/error/         3 rules (typed-error channel)
- effect/discriminants/ 2 rules (typed-union shape)

Also mirrors the new sub-family layout into docs/rules/effect/ so the
doc layout matches the source layout (per v0.0.8's principle that doc
layout mirrors source family folders).

src/rules/effect/index.ts now imports only from the 6 sub-family
index.ts files; src/rules/effect/README.md updated to point at the
sub-family READMEs.

909 tests pass. Lint clean (0 errors, 0 warnings).
