+++ docs/rules/effect/error/effect-promise.md
