+++ docs/rules/effect/error/no-effect-error-coalescing.md
