+++ docs/rules/effect/observability/annotate-without-span.md
