+++ docs/rules/effect/observability/handler-requires-span.md
