+++ docs/rules/effect/observability/logger-config-at-boot.md
