+++ docs/rules/effect/observability/no-console-in-effect.md
