+++ docs/rules/effect/observability/require-span-on-exported-effect.md
