+++ docs/rules/effect/runtime/no-promise-all-in-effect.md
