+++ docs/rules/effect/runtime/prefer-effect-platform.md
