+++ docs/rules/effect/schema/no-schema-type-cast.md
