+++ docs/rules/effect/scope/finalizer-requires-scope.md
