+++ docs/rules/effect/scope/runpromise-requires-scoped.md
