+++ src/rules/effect/discriminants/index.ts
@@ -1,0 +1,7 @@
+import eitherDiscriminant from "./either-discriminant.js";
+import tagDiscriminant from "./tag-discriminant.js";
+
+export const discriminantsRules = {
+  "either-discriminant": eitherDiscriminant,
+  "tag-discriminant": tagDiscriminant,
+} as const;
