+++ src/rules/effect/error/effect-error-erasure.test.ts
