+++ src/rules/effect/error/effect-error-erasure.ts
@@ -1,7 +1,7 @@
 import type { TSESTree } from "@typescript-eslint/utils";
 import { AST_NODE_TYPES } from "@typescript-eslint/utils";
-import { createRule } from "../../utils/create-rule.js";
-import { isNamedMemberCall } from "../../utils/ast-refinement/index.js";
+import { createRule } from "../../../utils/create-rule.js";
+import { isNamedMemberCall } from "../../../utils/ast-refinement/index.js";
 
 const GENERIC_ERROR_CTORS = new Set(["Error", "TypeError", "RangeError"]);
