+++ src/rules/effect/error/effect-promise.test.ts
