+++ src/rules/effect/error/no-effect-error-coalescing.test.ts
