+++ src/rules/effect/error/no-effect-error-coalescing.ts
@@ -1,7 +1,7 @@
 import type { TSESTree } from "@typescript-eslint/utils";
 import { AST_NODE_TYPES } from "@typescript-eslint/utils";
-import { createRule } from "../../utils/create-rule.js";
-import { isNamedMemberCall } from "../../utils/ast-refinement/index.js";
+import { createRule } from "../../../utils/create-rule.js";
+import { isNamedMemberCall } from "../../../utils/ast-refinement/index.js";
 
 const ERROR_COALESCING_METHODS = new Set(["catchAll", "catchAllCause", "mapError"]);
